Create a single method handle_posts in core/interaction.py. And call from handle_place, handle_hashtag.
